move redux hooks component
