+++ src/components/FormRegistration.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import { useAppDispatch, useAppSelector } from "../hooks";
+import { useAppDispatch, useAppSelector } from "../redux/hooks";
 import { addUser } from "../redux/slices/RegistrationSlices";
 import UserServices from "../services/UserServices";
 import {
